A program in Python which makes use of RSA to securely encrypt the content of emails.
A program in Python which makes use of RSA to securely encrypt the
content of emails. Implemented a database to keep track of all the
public certificates authorized by CA and successfully implemented a
method to encrypt and sign an email at sender’s side and a method to
decrypt the email, verify it for authenticity at receiver’s side.